@@ -1,11 +1,11 @@
 {
   "name": "functional-models-orm-dynamo",
-  "version": "2.0.4",
+  "version": "2.1.0",
   "description": "An implmentation of functional-models-orm for dynamodb.",
   "main": "index.js",
   "types": "index.d.ts",
   "scripts": {
-    "test": "mocha -r ts-node/register test/**/*.test.ts",
+    "test": "export TS_NODE_TRANSPILE_ONLY=true && export TS_NODE_PROJECT='./tsconfig.test.json' && mocha -r ts-node/register 'test/**/*.test.ts'",
     "test:coverage": "nyc npm run test",
     "feature-tests": "./node_modules/.bin/cucumber-js",
     "coverage": "nyc --all --reporter=lcov npm test",
@@ -49,32 +49,37 @@
     "report-dir": "coverage"
   },
   "dependencies": {
-    "functional-models": "^2.0.5",
+    "@aws-sdk/client-dynamodb": "^3.529.1",
+    "@aws-sdk/lib-dynamodb": "^3.529.1",
+    "functional-models": "^2.0.14",
     "functional-models-orm": "^2.0.17",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
-    "@istanbuljs/nyc-config-typescript": "^1.0.1",
-    "@types/chai": "^4.3.0",
-    "@types/lodash": "^4.14.177",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^16.11.7",
-    "@types/proxyquire": "^1.3.28",
-    "@types/sinon": "^10.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.7.0",
+    "@cucumber/cucumber": "^9.6.0",
+    "@istanbuljs/nyc-config-typescript": "^1.0.2",
+    "@types/chai": "^4.3.12",
+    "@types/lodash": "^4.17.0",
+    "@types/mocha": "^10.0.6",
+    "@types/node": "^20.11.27",
+    "@types/proxyquire": "^1.3.31",
+    "@types/sinon": "^17.0.3",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
     "babel-eslint": "^10.1.0",
-    "chai": "^4.3.0",
-    "cucumber": "^7.0.0-rc.0",
-    "eslint": "^7.19.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-functional": "^3.2.1",
-    "eslint-plugin-import": "^2.22.1",
-    "mocha": "^8.2.1",
+    "chai": "^4.2.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-functional": "^6.1.1",
+    "eslint-plugin-import": "^2.29.1",
+    "mocha": "^9.1.3",
     "nyc": "^15.1.0",
+    "prettier": "^3.2.5",
     "proxyquire": "^2.1.3",
-    "sinon": "^11.1.2",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.5.2"
+    "sinon": "^17.0.1",
+    "ts-mocha": "^10.0.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.2"
   },
   "homepage": "https://github.com/monolithst/functional-models-orm-dynamo#readme"
 }
