@@ -581,7 +581,4 @@ const RESERVED_KEYWORDS = {
 
 const SCAN_RETURN_THRESHOLD = 1000
 
-export {
-  RESERVED_KEYWORDS,
-  SCAN_RETURN_THRESHOLD,
-}
+export { RESERVED_KEYWORDS, SCAN_RETURN_THRESHOLD }
