@@ -4,6 +4,4 @@ const getTableNameForModel = (model: Model<any>) => {
   return model.getName().toLowerCase().replace('_', '-').replace(' ', '-')
 }
 
-export {
-  getTableNameForModel,
-}
+export { getTableNameForModel }
